moves duplicated code to testutil.ts
